requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
matplotlib
seaborn
requests
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
pypdf
duckduckgo-search
pydantic
openai
